pyproject.toml
[tool.poetry]
name = "nonebot-plugin-bwiki-navigator"
version = "0.2.0"
description = "A nonebot2 plugin version of bwiki official bot"
authors = ["XZhouQD <X.Zhou.QD@hotmail.com>"]
license = "AGPL v3"
readme = "README.md"
include = [
    "LICENSE",
]
homepage = "https://github.com/XZhouQD/nonebot-plugin-bwiki-navigator"
repository = "https://github.com/XZhouQD/nonebot-plugin-bwiki-navigator"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = "^2.0.0-beta.1"
httpx = { version = ">=0.20.0,<1.0.0", extras = ["http2"] }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
